Revert "Move channel and codec related files to common (#4225)"
This reverts commit 674a3f5869d3ae0e7ea3583a2d0942e184817903.
