+++ shell/platform/darwin/ios/framework/Source/FlutterStandardCodec_Internal.h
@@ -5,7 +5,7 @@
 #ifndef SHELL_PLATFORM_IOS_FRAMEWORK_SOURCE_FLUTTERSTANDARDCODECINTERNAL_H_
 #define SHELL_PLATFORM_IOS_FRAMEWORK_SOURCE_FLUTTERSTANDARDCODECINTERNAL_H_
 
-#include "flutter/shell/platform/darwin/common/FlutterCodecs.h"
+#include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterCodecs.h"
 
 typedef NS_ENUM(NSInteger, FlutterStandardField) {
   FlutterStandardFieldNil,
