+++ travis/licenses_golden/licenses_flutter
@@ -1162,20 +1162,20 @@ FILE: ../../../flutter/shell/platform/android/io/flutter/util/PathUtils.java
 FILE: ../../../flutter/shell/platform/android/io/flutter/util/Preconditions.java
 FILE: ../../../flutter/shell/platform/android/platform_view_android_jni.cc
 FILE: ../../../flutter/shell/platform/android/platform_view_android_jni.h
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterBinaryMessenger.h
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterChannels.h
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterChannels.mm
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterCodecs.h
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterCodecs.mm
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterStandardCodec.mm
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterStandardCodec_Internal.h
-FILE: ../../../flutter/shell/platform/darwin/common/flutter_codecs_unittest.mm
-FILE: ../../../flutter/shell/platform/darwin/common/flutter_standard_codec_unittest.mm
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterBinaryMessenger.h
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterChannels.h
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterCodecs.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterNavigationController.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterPlugin.h
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterChannels.mm
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterCodecs.mm
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterNavigationController.mm
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterStandardCodec.mm
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterStandardCodec_Internal.h
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/flutter_codecs_unittest.mm
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/flutter_main_ios.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/flutter_main_ios.mm
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/flutter_standard_codec_unittest.mm
 FILE: ../../../flutter/shell/platform/darwin/ios/ios_gl_context.h
 FILE: ../../../flutter/shell/platform/darwin/ios/ios_gl_context.mm
 FILE: ../../../flutter/shell/platform/darwin/ios/ios_surface.h
@@ -1427,7 +1427,6 @@ FILE: ../../../flutter/shell/platform/android/io/flutter/plugin/platform/Platfor
 FILE: ../../../flutter/shell/platform/android/io/flutter/view/VsyncWaiter.java
 FILE: ../../../flutter/shell/platform/android/vsync_waiter_android.cc
 FILE: ../../../flutter/shell/platform/android/vsync_waiter_android.h
-FILE: ../../../flutter/shell/platform/darwin/common/FlutterMacros.h
 FILE: ../../../flutter/shell/platform/darwin/common/buffer_conversions.mm
 FILE: ../../../flutter/shell/platform/darwin/common/process_info_mac.cc
 FILE: ../../../flutter/shell/platform/darwin/common/process_info_mac.h
@@ -1435,6 +1434,7 @@ FILE: ../../../flutter/shell/platform/darwin/desktop/vsync_waiter_mac.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/Flutter.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterAppDelegate.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterDartProject.h
+FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterMacros.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Headers/FlutterViewController.h
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterDartProject.mm
 FILE: ../../../flutter/shell/platform/darwin/ios/framework/Source/FlutterDartProject_Internal.h
